#149- Move next_path() to fs.py
