+++ tests/test_fs.py
@@ -1,6 +1,6 @@
-"""`pytest` tests for utils/logging.py"""
+"""`pytest` tests for utils/fs.py"""
 
-from benchcab.utils.logging import next_path
+from benchcab.utils.fs import next_path
 from .common import MOCK_CWD
